requirements.txt
discord.py==2.3.2
web3==6.15.1
python-dotenv==1.0.1
aiohttp==3.9.3
eth-abi==4.2.1
eth-account==0.10.0
eth-utils==3.0.0 